Remove console logging of exception messages
The code change removes the line that prints the exception message to the console within the catch block of a try-catch-finally statement. Specifically, the line `Console.WriteLine(ex.Message);` has been deleted. This means that if an exception occurs, the message will no longer be printed to the console, but the method will still return `false`.